# docs/openapi.yaml
openapi: 3.0.4
info:
  title: The Crazy Unicorn Shoppe
  version: 1.0.0
  description: >-
    This is an e-commerce full-stack app made by Steve Majewski for practice,
    the content is intended to be silly and far from serious, but it will be
    fully functional as far as creating accounts and logging in with hashed
    passwords. The data will be stored in a Postgres Database.
servers:
  - url: www.localhost:4700
    variables: {}
...
paths:
...
  /cart:
    summary: This path will add an item to the cart.
    post:
      summary: add cart item
      description: >-
        This operation will take a product and a quantity and add the item to
        the users active cart.
      operationId: post_cart
      requestBody:
        description: An item to add to the users cart.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cart_item'
      responses:
        '200':
          description: OK
          content:
            text/plain; charset=utf-8:
              examples:
                post cart item success:
                  value: |-
                    {
                      "cart_id": 38,
                      "product_id": 4,
                      "item_qty": 3
                    }
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - cart
    get:
      summary: get the cart items
      description: This endpoint will return all of the cart items for the current user.
      operationId: get_cart
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                get cart items success:
                  value: |-
                    [
                      {
                        "product_name": "Unicorn Pants",
                        "item_qty": 3
                      },
                      {
                        "product_name": "Unicorn Shirt",
                        "item_qty": 1
                      },
                      {
                        "product_name": "Unicorn Vest",
                        "item_qty": 2
                      }
                    ]
            text/plain; charset=utf-8:
              examples:
                Empty Cart:
                  value: That's one empty cart you got there!
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - cart
    put:
      summary: change cart item quantity
      description: This operation will change the quantity of a cart item.
      operationId: put_cart
      requestBody:
        description: Updates a cart item quantity.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Update_cart_item_qty'
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                update success:
                  value: |-
                    {
                      "cart_id": 39,
                      "product_id": 4,
                      "item_qty": 2
                    }
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - cart
  /cart/checkout:
    summary: will handle the checkout tasks
    get:
      summary: get data to checkout
      description: >-
        This operation will return the data needed to assemble and submit an
        order.
      operationId: get_cart_checkout
      responses:
        '200':
          description: Default error sample response
          content:
            application/json:
              examples:
                get checkout data success:
                  value: |-
                    [{
                      "cart_id": 477,
                      "user_id": 47,
                      "product_name": "Unicorn Slippers",
                      "item_qty": 2,
                      "product_price": 20
                    },
                    {
                      "cart_id": 477,
                      "user_id": 47,
                      "product_name": "Unicorn Car",
                      "item_qty": 1,
                      "product_price": 2000
                    },
                    {
                      "cart_id": 477,
                      "user_id": 47,
                      "product_name": "Unicorn Shirt",
                      "item_qty": 1,
                      "product_price": 20
                    }]
            text/plain; charset=utf-8:
              examples:
                empty cart:
                  value: Thats one empty cart you got there!
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - cart
    post:
      summary: accepts and runs payment, adjust inventory
      description: >-
        This operation will pretend to run the users payment method and return
        ok on success, it is setup in code to return a failed payment at
        random(roughly 1 in 25 is the chance of a failed payment). This will
        also be the operation that will handle adjusting the product inventory
        in the database.
      operationId: post_cart_checkout
      responses:
        '200':
          description: OK
        '400':
          description: payment rejected
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: The payment was rejected, please try another method.
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - cart
  /cart/{id}:
    summary: update quantity for a cart item
    delete:
      summary: remove a cart item
      description: >-
        This operation will take a product id as a url parameter and remove the
        item from the users cart.
      operationId: delete_cart_item
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the product we seek to remove.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain; charset=utf-8:
              examples:
                delete item success:
                  value: The item has successfully been removed from the cart.
        '401':
          description: not authorized
          content:
            text/plain; charset=utf-8:
              examples:
                not authorized:
                  value: You must be logged in to perform this action.
        '404':
          description: no active cart
          content:
            text/plain; charset=utf-8:
              examples:
                user has no active cart:
                  value: Add an item to create a cart.
                users cart does not have product:
                  value: This user does not have that product in their cart.
        '500':
          description: internal server error
          content:
            text/plain; charset=utf-8:
              examples:
                server error:
                  value: Something went wrong on our end, please try again.
      tags:
        - cart
...
components:
  schemas:
...
    Cart_item:
      type: object
      properties:
        product_id:
          type: integer
          format: int32
          example: 2
        item_qty:
          type: integer
          format: int32
          example: 5
    Update_cart_item_qty:
      type: object
      properties:
        product_id:
          type: integer
          format: int32
          example: 3
        item_qty:
          type: integer
          format: int32
          example: 2
...